requirements.txt
ipython[all]>=3.2.0
numpy>=1.9.2
scipy>=0.15.1
scikit-learn>=0.18
matplotlib>=1.4.3
pandas>=0.19
nltk
# tensorflow-macos
keras==2.12.0
torch
pydantic~=1.10.6
strictyaml~=1.7.3
pytest~=7.2.2
numpy~=1.24.2
setuptools==67.6.0
pandas~=1.5.3
joblib~=1.2.0
scikit-learn~=1.2.2
numpy>=1.21.0,<2.0.0
pandas>=1.3.5,<2.0.0
pydantic>=1.8.1,<2.0.0
scikit-learn>=1.1.3,<2.0.0
strictyaml>=1.3.2,<2.0.0
ruamel.yaml>=0.16.12,<1.0.0
feature-engine>=1.0.2,<2.0.0
joblib>=1.0.1,<2.0.0
matplotlib>=3.4.3,<4.0.0
scipy>=1.7.1,<2.0.0
statsmodels>=0.12.2,<1.0.0
fastapi
uvicorn
nltk~=3.8.1
PyYAML~=6.0